InstServer M2M: fix validation in InstServerForm
The validation code was relying on each server having just one institution.

With the M2M mapping of Institutions and Servers, change the logic in
clean_ertype to iterate over all institutions and check the server ertype
complies with all of the institutions / their ertype values.

For other server values required for an IdP (auth_port, acct_port, pkt_type),
check them against this server er_type.

When fetching the ertype of this server, check it has been set to avoid an
embarassing KeyError.