Django 1.8: add new generated initial migrations
Generated with:

    ./manage.py makemigrations --name initial

(__init__.py created manually)

Due to the structure of the accounts tables, there is more than one initial migration.

Note: this also includes the migrations generated for accounts.User, the custom
user model replacing django.contrib.auth.User.

Note that accounts.User is also a swappable model (same as django.contrib.auth.User).

And so it is important to also include the swappable property in the generated
migrations. (Without the property also being set in the generated migrations,
the migrations framework does not consider the class swapped out and attempts
to run the migrations for this class even when e.g. auth.User is used, clashing
with the migrations for auth.User and breaking the database initialization).