Django1.8: Custom User Model: port auth migrations
Port django.contrib.auth migrations affecting the user Model over to our Custom
User Model accounts.user as
0003_alter_user_email_max_length and
0004_alter_user_last_login_null.

Manually modify 0001_initial to pretend these changes haven't been done yet to
make the migrations framework actually apply these migrations.

Detailed explanation and reasoning:

In Django 1.7, South was dropped to use Django migrations instead.

Any DB history prior to that gets squashed into a new initial migration in the
new system.

In Django 1.8, with the new Django migrations, auth.User evolved to extend the
email address (from 75 to 254) and also adds blank=True to last_login.
For auth.User, these are captured in stand-alone migrations.

When replacing auth.User with account.User, it might be tempting to create just
one initial migration with the final version of the model.

However, this would not work for existing databases. Because we are at the same
time switching from South to Django migrations, we need to fake history on what
was done. Because the tables already exist, we need to run

    ./manage.py migrate --fake-initial

to fake the initial migration - and in some cases, also fake subsequent
migrations. E.g., because the edumanage package initial state is spread over
multiple migration steps, we do it with full:

    ./manage.py migrate --fake-initial --fake edumanage

and likewise for account:

    ./manage.py migrate --fake-initial --fake accounts 0002_initial

But, here we explicitly state that accounts should have the migration history
faked only up to 0002_initial.

At that point, we get the original tables created with Django 1.4 and with
older versions of email and last_login - and this is when we actually apply the
additional two migrations included in this commit:

    ./manage.py migrate --fake-initial

which lets migrate go wild and apply any unapplied migrations (but still faking initial ones).