parse_institution_xml: chg pk=1 to NRO_COUNTRY_CODE
In the parse_institution_xml management command, use settings.NRO_COUNTRY_CODE
instead of hard-coded pk=1.  Otherwise, the management command fails when the
primary key of the Realm object is not one.  (E.g., if the object was deleted
and re-created).

The value settings.NRO_COUNTRY_CODE is already used as a key into the Realm
model in edumanage/views.py in get_nro_name.

See also discussion in #18.