edumanage/context_processors: remove NRO_REALM
Remove the NRO_REALM context parameter definition in edumanage/context_processors.py

This got introduced partly accidentally in f7e106c as a result of discussion on #17.

However, this causes all page loads to break with a DoesNotExist exception if
the Realm object does not exist yet.

As the NRO_REALM context parameter is not used yet, the safest approach is to
remove it.