Fix custom User model: username MaxLengthValidator also needs patching
In the context of Django 1.8 migration, longerusername was replaced
with a custom User model. Like the former solution, this also patches
the username field to increase max_length. Alternatively one would
need to add a different field, since field name hiding is not
permitted[1], and have USERNAME_FIELD[2] point to it. But this
involves code duplication which seems unnecessary for such a small
change.

However the patching approach also has issues: It is applied after
CharField is instantiated, which means some things are setup before
max_length is adjusted as intended. MaxLengthValidator is such a case,
so it needs patching too. This is also the approach taken[3] by
longerusername.

This issue is evident in the Django admin interface, where editing any
user field and saving leads to a validation error if the username
exceeds the default max_length (30 characters).

Django documentation[4] suggests that UserCreationForm and UserChangeForm
ModelForms used in UserAdmin must also be overridden, so as to have their
Meta class point to the custom User model. This is again the approach[5]
in longerusername (for older Django versions though, no custom User
model involved). In practice our use case apparently does not require
this (on the contrary actually overriding the forms without adjusting
MaxLengthValidator does not have any effect whatsoever).

[1] https://docs.djangoproject.com/en/1.8/topics/db/models/#field-name-hiding-is-not-permitted
[2] https://docs.djangoproject.com/en/1.8/topics/auth/customizing/#django.contrib.auth.models.CustomUser.USERNAME_FIELD
[3] https://github.com/madssj/django-longer-username-and-email/blob/master/longerusernameandemail/models.py#L23
[4] https://docs.djangoproject.com/en/1.8/topics/auth/customizing/#custom-users-and-the-built-in-auth-forms
[5] https://github.com/madssj/django-longer-username-and-email/blob/master/longerusernameandemail/forms.py