Strings in local_settings.py may not be marked for translation; provide a workaround, only for templates, with in-place translation
After local_settings.py was introduced, we allowed for strings therein
to be marked for translation with ugettext_lazy(); however that does not
work in pricinple, since they should not be committed to the source
language file.

We now deprecate this and provide a partial workaround, only for
templates, with in-place translation through a dictionary keyed by
language code.