Yet another fix for in-place translation, thanks gettext
Django runs xgettext (through makemessages) with a number of keyword
options, without however resetting default keywords. In every file
xgettext scans for translatable strings, it looks for _() as well as
other keywords. For Python files this happens regardless whether a
gettext function has been imported as _ or not!

It is thus important that such a symbol is *only* ever used for a
gettext function that should mark strings for translation!

Therefore we need to import djnro.lldict.LazyLangDict by a different
name.