fix: mark CSRF cookies as secure
For security, they should be marked as secure (we already mark session cookies as secure).

And Firefox gives warnings:

Cookie “csrftoken” will be soon rejected because it has the “SameSite” attribute
set to “None” or an invalid value, without the “secure” attribute.
To know more about the “SameSite“ attribute, read
https://developer.mozilla.org/docs/Web/HTTP/Headers/Set-Cookie/SameSite