fix: edumanage/views/cat_user_api_proxy: 2to3 fix for CT handling
The code in cat_user_api_proxy changes Content-Type from text/html to application/json
if it detects the payload returned is JSON - starts with '[' or '{'.

However, this code was missed in the Python3 overhaul.

In Python Requests, the type of response.content is bytes.

The existing comparison: r.content[0] in ['{', '[']
works in Python2, where strings and bytes are not distinguished.

However, in Python3, r.content[0] returns just the value of the first byte (91),
which does not match up with single-character unicode strings.

Fix this to work in both Python2 and Python3 by:
* explicitly specifying the characters the value is compared to as bytes
* using a 1 element slice of r.content instead of a single index to avoid getting the byte value directly

So the new check, compatible with both Python2 and Python3, is: r.content[0:1] in [b'{', b'[']