new: edumanage/template: add link to GEANT Location codes
On the Institution Edit and Service LoocationEdit pages, add a link
to wiki page maintained by GEANT listing the location codes
that can be used for the Institution Type.

Make this text translatable.

Link to the page via Confluence tiny URL - which is equivalent to a page ID,
but redirects to readable URL with page title.